# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactResponse": {
        "properties": {
          "artifact_type": {
            "title": "Artifact Type",
            "type": "string"
          },
          "checksum": {
            "title": "Checksum",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "storage_uri": {
            "title": "Storage Uri",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "artifact_type",
          "storage_uri",
          "checksum",
          "metadata",
          "created_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
      "AuditEventResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mission_id": {
            "format": "uuid",
            "title": "Mission Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "id",
          "event_type",
          "actor",
          "mission_id",
          "run_id",
          "task_id",
          "status",
          "occurred_at",
          "metadata"
        ],
        "title": "AuditEventResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeDocumentCreateRequest": {
        "properties": {
          "source_metadata": {
            "additionalProperties": true,
            "title": "Source Metadata",
            "type": "object"
          },
          "source_uri": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Uri",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "source_uri",
          "text"
        ],
        "title": "KnowledgeDocumentCreateRequest",
        "type": "object"
      },
      "KnowledgeDocumentResponse": {
        "properties": {
          "checksum": {
            "title": "Checksum",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingestion_status": {
            "title": "Ingestion Status",
            "type": "string"
          },
          "source_uri": {
            "title": "Source Uri",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_uri",
          "ingestion_status",
          "checksum",
          "created_at"
        ],
        "title": "KnowledgeDocumentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AegisOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/artifacts/{artifact_id}/content": {
      "get": {
        "description": "Serve an artifact's raw file content (e.g. the report JSON, with its\nembedded slide deck) so the frontend can render or download it --\n``storage_uri`` alone is a ``file://`` URI a browser can't fetch.",
        "operationId": "get_artifact_content_route_api_v1_workspaces__workspace_id__artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Identity-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Identity-Subject"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Display-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Display-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Content Route",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/audit-events": {
      "get": {
        "operationId": "search_audit_events_route_api_v1_workspaces__workspace_id__audit_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Identity-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Identity-Subject"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Display-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Display-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response Search Audit Events Route Api V1 Workspaces  Workspace Id  Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Audit Events Route",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/knowledge/documents": {
      "get": {
        "operationId": "list_documents_route_api_v1_workspaces__workspace_id__knowledge_documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Identity-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Identity-Subject"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Display-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Display-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDocumentResponse"
                  },
                  "title": "Response List Documents Route Api V1 Workspaces  Workspace Id  Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents Route",
        "tags": [
          "knowledge"
        ]
      },
      "post": {
        "operationId": "ingest_document_route_api_v1_workspaces__workspace_id__knowledge_documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Identity-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Identity-Subject"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Display-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Display-Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Document Route",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/missions/{mission_id}/artifacts": {
      "get": {
        "operationId": "list_artifacts_route_api_v1_workspaces__workspace_id__missions__mission_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Identity-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Identity-Subject"
            }
          },
          {
            "in": "header",
            "name": "X-Aegis-Display-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aegis-Display-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactResponse"
                  },
                  "title": "Response List Artifacts Route Api V1 Workspaces  Workspace Id  Missions  Mission Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artifacts Route",
        "tags": [
          "artifacts"
        ]
      }
    },
...
  }
}